+++ changelog.md
@@ -1,5 +1,13 @@
 # LiteMigrator Change Log
 
+## v0.9.0
+
+Breaking Changes:
+
+* [NEW] Renamed class `LiteMigration` to `Migrator`
+* [FIX] BaseAssembly object is used instead of BaseAssemblyFile (string)
+  * This fixes loading migrations on Android
+
 ## v0.8.0
 
 Constructor and namespace cleanup.